PolyGerrit: Add bulk deleting of branches

+++ polygerrit-ui/app/elements/admin/gr-repo-detail-list/gr-repo-detail-list.js
@@ -73,7 +73,7 @@
         value: true,
       },
       _filter: String,
-      _refName: String,
+      _refName: Object,
       _hasNewItemName: Boolean,
       _isEditing: Boolean,
       _revisedRef: String,
@@ -92,6 +92,10 @@
 
     _paramsChanged(params) {
       if (!params || !params.repo) { return; }
+
+      this._refName = {
+        item: [],
+      };
 
       this._repo = params.repo;
 
@@ -200,7 +204,24 @@
       }
     },
 
+    _handleTargetTap(e) {
+      let checkbox = Polymer.dom(e.target).querySelector('input');
+      if (checkbox) {
+        checkbox.click();
+      } else {
+        // The target is the checkbox itself.
+        checkbox = Polymer.dom(e).rootTarget;
+      }
+      const stripRef = this._stripRefs(checkbox.name, this.detailType);
+      if (checkbox.checked) {
+        this._refName.item.push(stripRef);
+      } else {
+        this._refName.item.splice(this._refName.item.indexOf(stripRef), 1);
+      }
+    },
+
     _handleDeleteItemConfirm() {
+      if (!this._refName || !this._refName.item.length) { return; }
       this.$.overlay.close();
       if (this.detailType === DETAIL_TYPES.BRANCHES) {
         return this.$.restAPI.deleteRepoBranches(this._repo,
@@ -230,9 +251,6 @@
     },
 
     _handleDeleteItem(e) {
-      const name = this._stripRefs(e.model.get('item.ref'), this.detailType);
-      if (!name) { return; }
-      this._refName = name;
       this.$.overlay.open();
     },
 
@@ -266,6 +284,14 @@
 
     _computeHideTagger(tagger) {
       return tagger ? '' : 'hide';
+    },
+
+    _deleteButtonCheckArrayLength(name) {
+      if (name.item.length) {
+        return true;
+      }
+
+      return false;
     },
   });
 })();
